# docs/swagger.yaml
definitions:
...
  requestmodels.DishUpdateReq:
    properties:
      allergen_information:
        maxLength: 50
        type: string
      availability:
        type: boolean
      calories:
        type: integer
      carbohydrates:
        type: integer
      category_id:
        type: integer
      cuisine_type:
        minLength: 2
        type: string
      description:
        minLength: 5
        type: string
      dietary_information:
        type: string
      fat:
        type: integer
      mrp:
        type: number
      name:
        minLength: 2
        type: string
      portion_size:
        type: string
      preparation_time:
        maxLength: 15
        type: string
      price:
        type: number
      promotion_discount:
        type: integer
      protein:
        type: integer
      quantity:
        type: integer
      recommended_pairings:
        maxLength: 50
        type: string
      restaurantId:
        type: string
      special_features:
        maxLength: 50
        type: string
      spice_level:
        maxLength: 20
        type: string
      story_origin:
        maxLength: 300
        type: string
    required:
    - availability
    - category_id
    - cuisine_type
    - description
    - mrp
    - name
    - quantity
    - restaurantId
    type: object
...
  requestmodels.OtpVerification:
    properties:
      otp:
        type: string
    required:
    - otp
    type: object
...
  requestmodels.UserLoginReq:
    properties:
      password:
        minLength: 4
        type: string
      phone:
        minLength: 10
        type: string
    required:
    - password
    - phone
    type: object
  requestmodels.UserSignUpReq:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        minLength: 3
        type: string
      lastName:
        minLength: 1
        type: string
      password:
        minLength: 3
        type: string
      phone:
        type: string
    required:
    - confirmPassword
    - email
    - firstName
    - lastName
    - password
    - phone
    type: object
...
  responsemodels.Response:
    properties:
      after execution: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: Online Veg Food Delivery.
  title: Go + Gin Veg*n.
  version: "1.0"
paths:
...
  /dishes:
    get:
      consumes:
      - application/json
      description: Retrieves all dishes.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: GetAllDishesForUser
      tags:
      - Dish
  /dishes/{categoryid}:
    get:
      consumes:
      - application/json
      description: Retrieves dishes for a specific category.
      parameters:
      - description: The ID of the category to fetch dishes for.
        in: path
        name: categoryid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: FetchDishesByCategoryId
      tags:
      - Dish
  /dishes/{dishid}:
    get:
      consumes:
      - application/json
      description: Retrieves details for a specific dish.
      parameters:
      - description: The ID of the dish to fetch.
        in: path
        name: dishid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: FetchDishWithId
      tags:
      - Dish
  /login/:
    post:
      consumes:
      - application/json
      description: User can login using this handler.
      parameters:
      - description: Login Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: UserLogin
      tags:
      - User
...
  /profile:
    get:
      consumes:
      - application/json
      description: Retrieves the profile information for a specific user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GetUserProfile
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Edits a user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: EditUserProfile
      tags:
      - User
...
  /restaurant/dish:
    get:
      consumes:
      - application/json
      description: Fetch all dishes for the restaurant specified in the request context
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Fetch all dishes for a restaurant
      tags:
      - RestauratDishManagement
    post:
      consumes:
      - application/json
      description: Add a new dish for the restaurant specified in the request context
      parameters:
      - description: Name of the dish
        in: formData
        name: name
        required: true
        type: string
      - description: Category ID of the dish
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Description of the dish
        in: formData
        name: description
        required: true
        type: string
      - description: Cuisine type of the dish
        in: formData
        name: cuisine_type
        required: true
        type: string
      - description: MRP (Maximum Retail Price) of the dish
        in: formData
        name: mrp
        required: true
        type: number
      - description: Portion size of the dish
        in: formData
        name: portion_size
        type: string
      - description: Dietary information of the dish
        in: formData
        name: dietary_information
        type: string
      - description: Calories in the dish
        in: formData
        name: calories
        type: integer
      - description: Protein content in the dish
        in: formData
        name: protein
        type: integer
      - description: Carbohydrates content in the dish
        in: formData
        name: carbohydrates
        type: integer
      - description: Fat content in the dish
        in: formData
        name: fat
        type: integer
      - description: Spice level of the dish
        in: formData
        name: spice_level
        type: string
      - description: Allergen information of the dish
        in: formData
        name: allergen_information
        type: string
      - description: Recommended pairings for the dish
        in: formData
        name: recommended_pairings
        type: string
      - description: Special features of the dish
        in: formData
        name: special_features
        type: string
      - description: Image file of the dish
        in: formData
        name: image
        required: true
        type: file
      - description: Preparation time for the dish
        in: formData
        name: preparation_time
        type: string
      - description: Promotion discount for the dish
        in: formData
        name: promotion_discount
        type: integer
      - description: Story origin of the dish
        in: formData
        name: story_origin
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Add a new dish
      tags:
      - RestauratDishManagement
  /restaurant/dish/{dishid}:
    patch:
      consumes:
      - application/json
      description: Update details for the dish specified by dishid
      parameters:
      - description: ID of the dish to update.
        in: path
        name: dishid
        required: true
        type: integer
      - description: Updated dish information
        in: body
        name: dishData
        required: true
        schema:
          $ref: '#/definitions/requestmodels.DishUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: Update details for a dish
      tags:
      - RestauratDishManagement
...
  /signup/:
    post:
      consumes:
      - application/json
      description: User can signup using this handler
      parameters:
      - description: User Sign-Up Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.UserSignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      summary: UserSignUp
      tags:
      - User
  /verify/:
    post:
      consumes:
      - application/json
      description: User can verify the OTP which is generated after successful signup
        request.
      parameters:
      - description: OTP generated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.OtpVerification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - OtpTempTokenAuth: []
      summary: UserOTPVerication
      tags:
      - User
securityDefinitions:
  AdminRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  OtpTempTokenAuth:
    in: header
    name: Authorizations
    type: apiKey
  RestaurantAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  RestaurantRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  UserAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  UserRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
swagger: "2.0"